% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{b_12}
\alias{b_12}
\title{b(P1, P2, r) function}
\usage{
b_12(r, poly_1, poly_2)
}
\arguments{
\item{r}{\code{numeric scalar} a distance}

\item{poly_1}{\code{SpatialPolygons} - Polygon pattern 1}

\item{poly_2}{\code{SpatialPolygons} - Polygon pattern 2}
}
\value{
\code{numeric scaler}
}
\description{
Internal use
}
