% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_psam}
\alias{calc_psam}
\title{Calculates the psam}
\usage{
calc_psam(x, method = "rnd_poly")
}
\arguments{
\item{x}{a distance matrix}

\item{method}{method to deal with broken polygons}
}
\value{
scalar psam
}
\description{
Computes the PSAM based on a distance matrix based on a method
to deal with broken polygons (represented by duplicated rows)
}
