% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{covariog_aux}
\alias{covariog_aux}
\title{Auxiliar function to calculate the Covariogram from Lavancier et al. 2019}
\usage{
covariog_aux(h, dists, resp, tol, n_xy)
}
\arguments{
\item{h}{\code{numeric scalar}}

\item{dists}{\code{numeric vector}}

\item{resp}{\code{numeric vector}}

\item{tol}{\code{numeric scalar}}

\item{n_xy}{\code{numeric scalar}}
}
\value{
\code{numeric vector}
}
\description{
Auxiliar function to calculate \eqn{\hat{H_{12}}_i(d)}.
}
