% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tshift.R
\name{create_jump}
\alias{create_jump}
\title{Create jumps for random movements}
\usage{
create_jump(unique_bb)
}
\arguments{
\item{unique_bb}{a bbox shared between both "Polygon Patterns"}
}
\value{
a \code{sfc} object representing a random jump or shift.
}
\description{
Create jumps for random movements
}
\details{
This is an internal function.
}
\author{
Lucas Godoy
}
