% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h-fun.R
\name{fix_dist}
\alias{fix_dist}
\title{Fix distance matrix containing broken polygons}
\usage{
fix_dist(x, method = "rnd_poly")
}
\arguments{
\item{x}{distance matrix}

\item{method}{method used to fix. The options are "min", "max", "mean",
"rnd_poly", "rnd_dist", "min_norm", "max_norm", "hybrid", "hyb_center",
"hybrid_nc", "old_min"}
}
\value{
a distance matrix
}
\description{
fix a polygons' distance matrix based on a given method. This
function assumes the polygon that has been broken is represented by the
rows of the distance matrix.
}
