% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_test.R
\name{gof_mc}
\alias{gof_mc}
\title{Polygons Spatial Association Test - Global Envelope}
\usage{
gof_mc(
  obj_sp1,
  obj_sp2,
  n_sim = 499L,
  unique_bbox = NULL,
  alpha = 0.01,
  H = "L",
  ts = "SMAD",
  distances = NULL,
  fixed = FALSE,
  method = "hausdorff"
)
}
\arguments{
\item{obj_sp1}{an object from class \code{SpatialPolygons} or \code{SpatialPointsDataFrame}}

\item{obj_sp2}{an object from class \code{SpatialPolygons} or \code{SpatialPointsDataFrame}}

\item{n_sim}{an \code{integer} corresponding to the number of Monte Carlo simulations
for the test}

\item{unique_bbox}{a \code{matrix} \eqn{2 \times 2} corresponding to the boundary box
that contains both sets}

\item{alpha}{a \code{numeric} indicating the confidence level.}

\item{H}{a \code{character} indicating the function to be used. Possible entries are:
\code{'K'} or \code{'L'}.}

\item{ts}{a \code{character} associated to a test statistic. Inputs acepted:
\code{c('IM', 'MAD', 'SIM', 'SMAD', 'IMDQ', 'MADDQ')}.}

\item{distances}{a \code{numeric vector} indicating the distances to evaluate \eqn{H(d)}. If
\code{NULL} then the range considered goes from 5\% to 20\% of the max distance that can be
observed inside the \code{unique_bbox}.}

\item{fixed}{a \code{boolean} indicating if the first pattern should be fixed on the toroidal
shift or the first will be fixed in half of iterations and then the other one. \code{TRUE} or
\code{FALSE}, respectively.}

\item{method}{a \code{character} specifying which kind of distance will be used
to evalueate \eqn{H}. Also, there is an option of using areas. Options available:
\code{c('hausdorff', 'euclidean', 'area')}.}
}
\value{
a list from class \code{\link{gof_test}}, with values: \describe{
\item{p_value}{a \code{numeric} scalar giving the p-value of the test}
\item{mc_sample}{a \code{numeric} vector giving the test statistic for each of the Monte Carlo simulations}
\item{mc_funct}{a \code{matrix} where each line correspond to the function (\eqn{K} or \eqn{L}) estimated
for the Monte Carlo simulations}
\item{distances}{\code{numeric vector} containing the distances where mc_func were evaluated.}
\item{alpha}{a \code{numeric} scalar giving the significance level}
\item{rejects}{a \code{logical} scalar, TRUE if the null hypothesis is reject}
}
}
\description{
A Monte Carlo test to verify if two sets of polygons are
associated based in a global envelope of the functions
\eqn{K_{12}(d)} and \eqn{L_{12}(d)} using different test statistics.
}
