% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_func}
\alias{h_func}
\title{\eqn{H_{12}(d)}}
\usage{
h_func(
  obj_sp1,
  obj_sp2,
  unique_bbox = NULL,
  distances = NULL,
  method = "hausdorff",
  H = "L",
  correction = "none",
  ...
)
}
\arguments{
\item{obj_sp1}{an object from class \code{SpatialPolygons} or \code{SpatialPointsDataFrame}}

\item{obj_sp2}{an object from class \code{SpatialPolygons} or \code{SpatialPointsDataFrame}}

\item{unique_bbox}{a \code{matrix} \eqn{2 \times 2} corresponding to the boundary box
that contains both sets}

\item{distances}{a \code{numeric vector} indicating the distances to evaluate \eqn{H(d)}. If
\code{NULL} then the range considered goes from 5% to 20% of the max distance that can be
observed inside the \code{unique_bbox}.}

\item{method}{a \code{character} specifying which kind of distance will be used
to evalueate \eqn{H}. Also, there is an option of using areas. Options available:
\code{c('hausdorff', 'euclidean', 'area')}.}

\item{H}{a \code{character} indicating the function to be used. Possible entries are:
\code{'K'} or \code{'L'}.}

\item{correction}{a \code{character} giving the edge correction to be used. Possible
entries are \code{c('none', 'torus', 'guard', 'adjust')}.}

\item{...}{parameters for test statistics functions}
}
\value{
\code{numeric vector}.
}
\description{
Evaluate a given function in the inputted distances.
}
