% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_haus}
\alias{h_haus}
\title{H(d) - Hausdorff Distance}
\usage{
h_haus(obj_sp1, obj_sp2, distances, H, unique_bbox, ...)
}
\arguments{
\item{obj_sp1}{an object from class \code{SpatialPolygons} or \code{SpatialPointsDataFrame}}

\item{obj_sp2}{an object from class \code{SpatialPolygons} or \code{SpatialPointsDataFrame}}

\item{distances}{a \code{numeric vector} indicating the distances to evaluate \eqn{H(d)}. If
\code{NULL} then the range considered goes from 5% to 20% of the max distance that can be
observed inside the \code{unique_bbox}.}

\item{H}{a \code{character} indicating the function to be used. Possible entries are:
\code{'K'} or \code{'L'}.}

\item{unique_bbox}{a \code{matrix} \eqn{2 \times 2} corresponding to the boundary box
that contains both sets}

\item{...}{parameters for test statistics functions}
}
\value{
\code{numeric vector}.
}
\description{
Auxiliar function to \code{\link{h_func}}.
}
\details{
Internal use.
}
