% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h-fun.R
\name{hfun}
\alias{hfun}
\alias{h_func}
\alias{h_func.list}
\title{\eqn{h_{12}(t)} from polygons}
\usage{
h_func(
  p1,
  p2,
  hausdorff = TRUE,
  method = "rnd_poly",
  var_st = FALSE,
  dists = NULL
)

h_func.list(x, ...)
}
\arguments{
\item{p1}{sf object}

\item{p2}{sf object}

\item{hausdorff}{logical parameter indicating whether the Hausdorff distance
should be used}

\item{method}{method to deal with broken polygons}

\item{var_st}{logical scalar indicating if the L function should be used
instead}

\item{dists}{vector of distances to compute \eqn{h_{12}(t)}.}

\item{x}{a list with two sf objects.}

\item{...}{Parameters to be used with \code{h_func} when inputting a list.}
}
\value{
a numeric vector
}
\description{
Computes the \eqn{h_{12}} (K or L) based on a distance matrix
based on a method
}
