% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h-fun.R
\name{iadist}
\alias{iadist}
\title{ID aware distance matrix}
\usage{
iadist(p1, p2, hausdorff = TRUE, method = "rnd_poly")
}
\arguments{
\item{p1}{a \code{sf} object containing one column specifying the objects
id.}

\item{p2}{a \code{sf} object containing one column specifying the objects
id.}

\item{hausdorff}{\code{logical} scalar indicating whether the Hausdorff
distance should be used.}

\item{method}{method for "fixing" the distance matrix.}
}
\value{
a distance matrix.
}
\description{
Distance between polygons accounting for toroidal shift.
}
\author{
Lucas Godoy
}
