% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{im}
\alias{im}
\title{Integram Measure}
\usage{
im(x, h = 1)
}
\arguments{
\item{x}{\code{numeric matrix}}

\item{h}{\code{numeric}}
}
\value{
\code{numeric vector}
}
\description{
Integram Measure
}
