% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{im_ac}
\alias{im_ac}
\title{Integram Measure with Assimetry Correction}
\usage{
im_ac(x, h = 1)
}
\arguments{
\item{x}{\code{numeric matrix}}

\item{h}{\code{numeric}}
}
\value{
\code{numeric vector}
}
\description{
Integram Measure with Assimetry Correction
}
