% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_2.R
\name{kdist}
\alias{kdist}
\title{\eqn{K_{1,2}} - distance-based adaptation for polygons 2}
\usage{
kdist(obj_sp1, obj_sp2, r_min = NULL, r_max = NULL, by = NULL, bbox,
  sim = F, ...)
}
\arguments{
\item{obj_sp1}{a \code{SpatialPolygons} or \code{SpatialPointsDataFrame} object}

\item{obj_sp2}{a \code{SpatialPolygons} or \code{SpatialPointsDataFrame} object}

\item{r_min}{a \code{numeric} representing the mininmun distance to calculate
the function, \code{default = 0}.}

\item{r_max}{a \code{numeric} representing the maximum distance to calculate
the function \code{default = max(polygonds distance matrix)}.}

\item{by}{a \code{numeric} value that represents the how many units
between each \code{r} value.}

\item{bbox}{a \code{matrix} giving the region of study
entries are \code{c('none', 'torus', 'guard', 'adjust')}.}

\item{sim}{\code{boolean}}

\item{...}{parameters associated with edge correction.}
}
\value{
a \code{data.frame} corresponding to the estimated
 \eqn{K_{1,2}} function in the interval \eqn{[r_{min}, r_{max}]}.
}
\description{
test statistic for the MC test.
}
