% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{limits_to_sp}
\alias{limits_to_sp}
\title{Limits to Spatial}
\usage{
limits_to_sp(b_box)
}
\arguments{
\item{b_box}{Boundary box from class \code{matrix}}
}
\value{
an object from class \code{SpatialPolygons}
}
\description{
Internal function.
}
