% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{mad_ac}
\alias{mad_ac}
\title{Maximum Absolute Deviation with Assimetry Correction}
\usage{
mad_ac(x)
}
\arguments{
\item{x}{\code{numeric matrix}}
}
\value{
\code{numeric vector}
}
\description{
Maximum Absolute Deviation with Assimetry Correction
}
