% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_guard}
\alias{make_guard}
\title{Make Guard}
\usage{
make_guard(bbox, guard_perc = 0.125)
}
\arguments{
\item{bbox}{\code{numeric matrix}}

\item{guard_perc}{\code{numeric} between 0 and 1}
}
\value{
\code{numeric matrix}
}
\description{
Internal use
}
