% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_functions.R
\name{mc_iterations}
\alias{mc_iterations}
\title{psat_mc2 - auxiliar function}
\usage{
mc_iterations(obj1_shift, obj_sp2, niter, ts, args = NULL)
}
\arguments{
\item{obj1_shift}{object from class \code{SpatialPolygons} or \code{SpatialPoints}}

\item{obj_sp2}{object from class \code{SpatialPolygons} or \code{SpatialPoints}}

\item{niter}{\code{numeric} value indicating number of iterations}

\item{ts}{\code{string} giving the test statistic, the options are \code{c('psam', 'pf12', 'pk12')}}

\item{args}{\code{list} only necessary when the test statistic is different from psam}
}
\value{
a \code{numeric vector}.
}
\description{
Internal use.
}
