% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-stat.R
\name{mean_aux}
\alias{mean_aux}
\title{auxiliary mean}
\usage{
mean_aux(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector
}
\description{
aux function to calculate the mean of a vector when removing each of its
elements one by one.
}
\author{
Lucas Godoy
}
