% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mean_vec}
\alias{mean_vec}
\title{Vector of means}
\usage{
mean_vec(x)
}
\arguments{
\item{x}{\code{numeric vector}}
}
\value{
\code{numeric vector}
}
\description{
Auxiliar function to calculate \eqn{\hat{H_{12}}_i(d)}.
}
