% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pam}
\alias{pam}
\title{Polygons' Association Measure}
\usage{
pam(obj_sp1, obj_sp2)
}
\arguments{
\item{obj_sp1}{a \code{SpatialPolygons} or \code{SpatialPointsDataFrame} object}

\item{obj_sp2}{a \code{SpatialPolygons} or \code{SpatialPointsDataFrame} object}
}
\value{
a \code{scalar} \code{numeric}
}
\description{
test statistic for the MC test.
}
