% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/past.R
\docType{package}
\name{past}
\alias{past}
\alias{past-package}
\title{past: A package for testing polygons spatial association.}
\description{
The past package provides an important functons with methods to sumarise and
visualize it.
}
\section{Past functions and methods}{

\itemize{
  \item poly_mc
  \item summary
  \item plot
}
}

