% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat.R
\docType{package}
\name{pat}
\alias{pat}
\alias{pat-package}
\title{pat: A package for testing polygons/polygons or point/polygons pssociation.}
\description{
The pat package provides an important functons with methods to sumarise and
visualize it.
}
\section{Pat functions and methods}{

\itemize{
  \item poly_mc
  \item summary
  \item plot
}
}

