% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pf12}
\alias{pf12}
\title{F function adaptation for polygons}
\usage{
pf12(obj_sp1, obj_sp2, r_min = 0, r_max = NULL, by = 1)
}
\arguments{
\item{obj_sp1}{a \code{SpatialPolygons} or \code{SpatialPointsDataFrame} object}

\item{obj_sp2}{a \code{SpatialPolygons} or \code{SpatialPointsDataFrame} object}

\item{r_min}{a \code{numeric} representing the mininmun distance to calculate
the function, \code{default = 0}.}

\item{r_max}{a \code{numeric} representing the maximum distance to calculate
the function \code{default = max(polygonds distance matrix)}.}

\item{by}{a \code{numeric} value that represents the how many units
between each \code{r} value.}
}
\value{
a \code{numeric} \code{matrix} corresponding to the estimated
 \eqn{F_{1,2}} and \eqn{F_{2,1}} function in the interval \eqn{[r_{min}, r_{max}]}.
}
\description{
test statistic for the MC test.
}
