% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.pF}
\alias{plot.pF}
\alias{plot}
\title{Plot method for \code{\link{pF}}}
\usage{
\method{plot}{pF}(x, ...)
}
\arguments{
\item{x}{a \code{\link{pF}} object}

\item{...}{inherits from \code{plot}.}
}
\description{
Plot method for \code{\link{pF}}
}
