% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.psa_test}
\alias{plot.psa_test}
\alias{plot}
\title{PLot method for \code{\link{psa_test}}}
\usage{
\method{plot}{psa_test}(x, ...)
}
\arguments{
\item{x}{a \code{\link{psa_test}} object}

\item{...}{inherits from \code{plot}}
}
\description{
PLot method for \code{\link{psa_test}}
}
