% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{poly_rf}
\alias{poly_rf}
\title{Polygons' Random Shift}
\usage{
poly_rf(obj_sp, bbox_max)
}
\arguments{
\item{obj_sp}{object from class \code{SpatialPolygons}}

\item{bbox_max}{Boundary box from class \code{matrix}}
}
\value{
an object from class \code{SpatialPolygons} randomly translated
}
\description{
Polygons' Random Shift
}
