% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{poly_shift}
\alias{poly_shift}
\title{Create copies of a set of polygons}
\usage{
poly_shift(obj_sp, bbox_tot)
}
\arguments{
\item{obj_sp}{object from class \code{SpatialPolygons}}

\item{bbox_tot}{Boundary box from class \code{matrix}}
}
\value{
an object from class \code{SpatialPolygons}
}
\description{
Auxiliar function.
}
