% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{poly_touch}
\alias{poly_touch}
\title{Polygons that touch a bbox}
\usage{
poly_touch(x, bbox)
}
\arguments{
\item{objsp}{a \code{SpatialPolygon}}

\item{bbox_tot}{a \code{numeric matrix}}
}
\description{
Polygons that touch a bbox
}
