% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PolyDist}
\alias{PolyDist}
\title{Poly dist3 C}
\usage{
PolyDist(objsp1, objsp2)
}
\arguments{
\item{objsp1}{object from class \code{SpatialPolygons}}

\item{objsp2}{object from class \code{SpatialPolygons}}
}
\value{
an object from class \code{SpatialPolygons} randomly translated
}
\description{
Poly dist3 C
}
