% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tshift.R
\name{pre_ts}
\alias{pre_ts}
\title{Pre-TS}
\usage{
pre_ts(poly, bb = NULL, id_col = NULL)
}
\arguments{
\item{poly}{an object of class \code{sf} or \code{sfc}.}

\item{bb}{(optional) a unique bounding box.}

\item{id_col}{a \code{character} indicating the id column in \code{poly}.}
}
\value{
an \code{sf} with 8 additional rigid copies of \code{poly}.
}
\description{
Create rigid copies of a polygon. This function an auxilliary
function for the Toroidal Shift method
}
\author{
Lucas Godoy
}
