% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.mc_psa}
\alias{print.mc_psa}
\title{Print method for \code{\link{mc_psa}}}
\usage{
\method{print}{mc_psa}(x, ...)
}
\arguments{
\item{x}{A \code{\link{mc_psa}} object}

\item{...}{inherits from \code{print}.}
}
\description{
Print method for \code{\link{mc_psa}}
}
