% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.past_test}
\alias{print.past_test}
\alias{print}
\title{Print method for \code{\link{past_test}}}
\usage{
\method{print}{past_test}(x, ...)
}
\arguments{
\item{x}{A \code{\link{past_test}} object}

\item{...}{inherits from \code{print} and \code{format}}
}
\description{
Print method for \code{\link{past_test}}
}
