% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.pat_test}
\alias{print.pat_test}
\alias{print}
\title{Print method for \code{\link{pat_test}}}
\usage{
\method{print}{pat_test}(x, ...)
}
\arguments{
\item{x}{A \code{\link{pat_test}} object}

\item{...}{inherits from \code{print} and \code{format}}
}
\description{
Print method for \code{\link{pat_test}}
}
