% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.psa_test}
\alias{print.psa_test}
\alias{print}
\title{Print method for \code{\link{psa_test}}}
\usage{
\method{print}{psa_test}(x, ...)
}
\arguments{
\item{x}{A \code{\link{psa_test}} object}

\item{...}{inherits from \code{print} and \code{format}}
}
\description{
Print method for \code{\link{psa_test}}
}
