% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{psa_pf12}
\alias{psa_pf12}
\title{Class of the output from \code{\link{psat_mc}} when the test statistic is \code{\link{pf12}}}
\usage{
psa_pf12(x)
}
\arguments{
\item{x}{Output from \code{\link{psat_mc}}}
}
\description{
Internal use
}
