% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{psa_pk12}
\alias{psa_pk12}
\title{Class of the output from \code{\link{psat_mc}} when the test statistic is \code{\link{pk_dist12}}
or \code{\link{pk_area12}}}
\usage{
psa_pk12(x)
}
\arguments{
\item{x}{Output from \code{\link{psat_mc}}}
}
\description{
Internal use
}
