% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{psa_psam}
\alias{psa_psam}
\title{Class of the output from \code{\link{psat_mc}} when the test statistic is \code{\link{psam}}}
\usage{
psa_psam(x)
}
\arguments{
\item{x}{Output from \code{\link{psat_mc}}}
}
\description{
Internal use
}
