% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psam}
\alias{psam}
\title{Polygons' Spatial Association Measure}
\usage{
psam(obj_sp1, obj_sp2)
}
\arguments{
\item{obj_sp1}{a \code{SpatialPolygons} or \code{SpatialPointsDataFrame} object}

\item{obj_sp2}{a \code{SpatialPolygons} or \code{SpatialPointsDataFrame} object}
}
\value{
a \code{numeric} \code{scalar} with the polygons' spatial
 association measure for the corresponding objects.
}
\description{
test statistic for the MC test.
}
