% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_test.R
\name{psam_mc}
\alias{psam_mc}
\title{Polygons Spatial Association Test - PSAM}
\usage{
psam_mc(
  obj_sp1,
  obj_sp2,
  n_sim = 499L,
  unique_bbox = NULL,
  alpha = 0.05,
  alternative = "two_sided",
  fixed = FALSE,
  hausdorff = F,
  ...
)
}
\arguments{
\item{obj_sp1}{an object from class \code{SpatialPolygons} or \code{SpatialPointsDataFrame}}

\item{obj_sp2}{an object from class \code{SpatialPolygons} or \code{SpatialPointsDataFrame}}

\item{n_sim}{an \code{integer} corresponding to the number of Monte Carlo simulations
for the test}

\item{unique_bbox}{a \code{matrix} \eqn{2 \times 2} corresponding to the boundary box
that contains both sets}

\item{alpha}{a \code{numeric} indicating the confidence level}

\item{alternative}{a \code{character} indicating the alternative hypothesis, it can be: "two_sided",
"repulsion", or "attraction" if you interest is  only check if the sets are independent
          or not, if the two sets repulses each other, or if the two sets attracts each other,
          respectively.}

\item{fixed}{a \code{boolean} indicating if the first pattern should be fixed on the toroidal
shift or the first will be fixed in half of iterations and then the other one. \code{TRUE} or
\code{FALSE}, respectively.}

\item{hausdorff}{a \code{boolean}. If \code{TRUE}, then the Hausdorff distance is used.
Otherwise the Euclidean distance is used. Default is \code{FALSE}.}

\item{...}{parameters for test statistics functions}
}
\value{
a list from class \code{\link{psam_test}}, with values: \describe{
    \item{p_value}{a \code{numeric} scalar giving the p-value of the test}
    \item{mc_sample}{a \code{numeric} vector giving the test statistic for each of the Monte Carlo simulations}
    \item{alternative}{a \code{character} giving the alternative hypothesis}
    \item{alpha}{a \code{numeric} scalar giving the significance level}
    \item{rejects}{a \code{logical} scalar, TRUE if the null hypothesis is reject}
  }
}
\description{
A Monte Carlo test based on the test statistic \code{\link{psam}}, to verify if
two sets of polygons are associated.
}
