% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_test.R
\name{psat_mc}
\alias{psat_mc}
\title{Polygons Spatial Association Test}
\usage{
psat_mc(obj_sp1, obj_sp2, n_sim = 500L, unique_bbox = NULL,
  alpha = 0.01, ts = "psam", alternative = "two_sided",
  correction = "none", fixed = FALSE, r_min = NULL, r_max = NULL,
  by = NULL, ...)
}
\arguments{
\item{obj_sp1}{an object from class \code{SpatialPolygons} or \code{SpatialPointsDataFrame}}

\item{obj_sp2}{an object from class \code{SpatialPolygons} or \code{SpatialPointsDataFrame}}

\item{n_sim}{an \code{integer} corresponding to the number of Monte Carlo simulations
for the test}

\item{unique_bbox}{a \code{matrix} \eqn{2 \times 2} corresponding to the boundary box
that contains the both sets}

\item{alpha}{a \code{numeric} indicating the confidence level}

\item{ts}{a \code{character} indicating the test statistic used in the test, the options are
\code{c('psam', 'pk_dist12', 'pk_area12', 'pf12')}.}

\item{alternative}{a \code{character} indicating the alternative hypothesis, it can be: "two_sided",
"repulsion", or "attraction" if you interest is  only check if the sets are independent
          or not, if the two sets repulses each other, or if the two sets attracts each other,
          respectively.}

\item{correction}{a \code{character} giving the edge correction to be used. Possible
entries are \code{c('none', 'torus', 'guard', 'adjust')}.}

\item{fixed}{a \code{boolean} indicating if the first pattern should be fixed on the toroidal
shift or the first will be fixed in half of iterations and then the other one. \code{TRUE} or
\code{FALSE}, respectively.}

\item{r_min}{min distance to calculate \eqn{K_{1,2}} functions. Used only if ts != 'psam'}

\item{r_max}{max distance to calculate \eqn{K_{1,2}} functions. Used only if ts != 'psam'}

\item{by}{controls how many values between \code{r_min} and \code{r_max} will by used
to calculate \eqn{K_{1,2}} functions. Used only if ts != 'psam'}

\item{...}{parameters for test statistics functions}
}
\value{
a list from class \code{\link{psa_test}}, with values: \describe{
    \item{p_value}{a \code{numeric} scalar giving the p-value of the test}
    \item{sample_ts}{a \code{numeric} scalar giving the test statistic calculated in the original sample}
    \item{mc_ts}{a \code{numeric} vector giving the test statistic for each of the Monte Carlo simulations}
    \item{alternative}{a \code{character} giving the alternative hypothesis}
    \item{alpha}{a \code{numeric} scalar giving the significance level used on the test}
    \item{rejects}{a \code{logical} scalar, TRUE if the null hypothesis is reject}
  }
}
\description{
A Monte Carlo test to verify if two sets of polygons are
   associated. The difference bewtween this function and the function
   \code{psat_mc} is that here you can choose the test statistic.
}
