% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_test.R
\name{psat_mc2}
\alias{psat_mc2}
\title{Polygons Spatial Association Test (Old version)}
\usage{
psat_mc2(obj_sp1, obj_sp2, n_sim = 100L, unique_bbox = NULL,
  same_bbox = T, bbox_1 = NULL, bbox_2 = NULL, alpha = 0.05,
  alternative = "two_sided")
}
\arguments{
\item{obj_sp1}{an object from class \code{SpatialPolygons} or \code{SpatialPointsDataFrame}}

\item{obj_sp2}{an object from class \code{SpatialPolygons} or \code{SpatialPointsDataFrame}}

\item{n_sim}{an \code{integer} corresponding to the number of Monte Carlo simulations
for the test}

\item{unique_bbox}{a \code{matrix} \eqn{2 \times 2} corresponding to the boundary box
that contains the both sets}

\item{same_bbox}{a \code{boolean} - (desnecessario)}

\item{bbox_1}{a \code{matrix} \eqn{2 \times 2} corresponding to the boundary box
of the first spatial object}

\item{bbox_2}{a \code{matrix} \eqn{2 \times 2} corresponding to the boundary box
of the second spatial object}

\item{alpha}{a \code{numeric} indicating the confidence level}

\item{alternative}{a string indicating the alternative hypothesis, it can be: "independece",
"repulsion", or "attraction" if you interest is  only check if the sets are independent
          or not, if the two sets repulses each other, or if the two sets attracts each other,
          respectively.}
}
\value{
a list from class \code{\link{psa_test}}, with values: \describe{
    \item{p_value}{a \code{numeric} scalar giving the p-value of the test}
    \item{sample_ts}{a \code{numeric} scalar giving the test statistic calculated in the original sample}
    \item{mc_ts}{a \code{numeric} vector giving the test statistic for each of the Monte Carlo simulations}
    \item{alternative}{a \code{character} giving the alternative hypothesis}
    \item{alpha}{a \code{numeric} scalar giving the significance level used on the test}
    \item{rejects}{a \code{logical} scalar, TRUE if the null hypothesis is reject}
  }
}
\description{
A Monte Carlo test to verify if two sets of polygons are
   or not.
}
