% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_functions.R
\name{pt_rf}
\alias{pt_rf}
\title{Points' Random Shift}
\usage{
pt_rf(obj_sp, bbox_max)
}
\arguments{
\item{obj_sp}{object from class \code{SpatialPoints}}

\item{bbox_max}{Boundary box from class \code{matrix}}
}
\value{
an object from class \code{SpatialPoints} randomly translated
}
\description{
Points' Random Shift
}
