% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-stat.R
\name{s_im}
\alias{s_im}
\title{Studentized Integram Measure}
\usage{
s_im(x, h = 1)
}
\arguments{
\item{x}{\code{numeric matrix}}

\item{h}{\code{numeric}}
}
\value{
\code{numeric vector}
}
\description{
Studentized Integram Measure
}
