% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{s_mad}
\alias{s_mad}
\title{Studentized Maximum Absolute Deviation}
\usage{
s_mad(x)
}
\arguments{
\item{x}{\code{numeric matrix}}
}
\value{
\code{numeric vector}
}
\description{
Studentized Maximum Absolute Deviation
}
