% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shift_aux}
\alias{shift_aux}
\title{Auxiliar function}
\usage{
shift_aux(obj_sp, obj_sp2, obj_sp3, obj_sp4, n_poly, range_x, range_y)
}
\arguments{
\item{obj_sp}{a spatial object}

\item{obj_sp2}{a spatial object}

\item{obj_sp3}{a spatial object}

\item{obj_sp4}{a spatial object}

\item{n_poly}{an integer}

\item{range_x}{a numeric}

\item{range_y}{a numeric}
}
\description{
Auxiliar function
}
