% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shift_aux_point}
\alias{shift_aux_point}
\title{Auxiliar function to \code{poly_shift}}
\usage{
shift_aux_point(obj_sp, obj_sp2, obj_sp3, obj_sp4)
}
\arguments{
\item{obj_sp}{a spatial object}

\item{obj_sp2}{a spatial object}

\item{obj_sp3}{a spatial object}

\item{obj_sp4}{a spatial object}
}
\value{
a spatial object
}
\description{
Auxiliar function to \code{poly_shift}
}
