% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sp_ID_dist}
\alias{sp_ID_dist}
\title{Distance Matrix - ID}
\usage{
sp_ID_dist(obj_sp1, obj_sp2, m = NULL)
}
\arguments{
\item{obj_sp1}{a \code{SpatialPolygons} or \code{SpatialPointsDataFrame} object}

\item{obj_sp2}{a \code{SpatialPolygons} or \code{SpatialPointsDataFrame} object}

\item{m}{a \code{character} equal to \code{c('haus1', 'haus2')}.
If not supplied, euclidean distance is performed.}
}
\value{
a distance matrix
}
\description{
Calculate the distance between polygons (or points)
from different types considering their ID's.
}
