% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sp_ID_haus2}
\alias{sp_ID_haus2}
\title{Hausdorff Distance Matrix - ID (Method 2)}
\usage{
sp_ID_haus2(obj_sp1, obj_sp2)
}
\arguments{
\item{obj_sp1}{a \code{SpatialPolygons} or \code{SpatialPointsDataFrame} object}

\item{obj_sp2}{a \code{SpatialPolygons} or \code{SpatialPointsDataFrame} object}
}
\value{
a distance matrix
}
\description{
Calculate the hausdorff distance between polygons (or points)
             from different types considering their ID's.
}
