% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.mc_psa}
\alias{summary.mc_psa}
\title{Summary method for \code{\link{mc_psa}}}
\usage{
\method{summary}{mc_psa}(object, ...)
}
\arguments{
\item{object}{a \code{\link{mc_psa}} object}

\item{...}{inherits from \code{summary}.}
}
\description{
Summary method for \code{\link{mc_psa}}
}
