% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.past_test}
\alias{summary.past_test}
\alias{summary}
\title{Summary method for \code{\link{past_test}}}
\usage{
\method{summary}{past_test}(object, ...)
}
\arguments{
\item{object}{a \code{\link{past_test}} object}

\item{...}{inherits from \code{print}, \code{format}, and \code{summary}}
}
\description{
Summary method for \code{\link{past_test}}
}
