% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.psa_test}
\alias{summary.psa_test}
\alias{summary}
\title{Summary method for \code{\link{psa_test}}}
\usage{
\method{summary}{psa_test}(object, ...)
}
\arguments{
\item{object}{a \code{\link{psa_test}} object}

\item{...}{inherits from \code{print}, \code{format}, and \code{summary}}
}
\description{
Summary method for \code{\link{psa_test}}
}
