% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tpsa.R
\name{theme_tpsa}
\alias{theme_tpsa}
\title{\code{ggplo2} theme for \code{tpsa}}
\usage{
theme_tpsa(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
\code{ggplo2} theme for \code{tpsa}
}
