% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tshift.R
\name{toroidal_shift}
\alias{toroidal_shift}
\title{Toroidal Shift}
\usage{
toroidal_shift(x, y, shifted = FALSE, unique_bb = NULL)
}
\arguments{
\item{x}{a \code{sf} or \code{sfc} object. Its \code{geometry} may contain POLYGONS and/or
POINTS.}

\item{y}{a \code{sf} or \code{sfc} object. Its \code{geometry} may contain POLYGONS and/or
POINTS.}

\item{shifted}{\code{logical} indicating whether \code{x} has been "shifted". This
parameter is mainly for internal use and testing.}

\item{unique_bb}{a bbox shared between both "Polygon Patterns"}
}
\value{
a list
}
\description{
Toroidal Shift
}
\author{
Lucas Godoy
}
