% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{torus_corr}
\alias{torus_corr}
\title{Toroidal edge correction}
\usage{
torus_corr(objsp, bbox_tot = NULL)
}
\arguments{
\item{objsp}{a \code{SpatialPolygon}}

\item{bbox_tot}{a \code{numeric matrix}}
}
\description{
Toroidal edge correction
}
