% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapo.R
\docType{package}
\name{tpsa}
\alias{sapo}
\alias{sapo-package}
\alias{tpsa}
\title{sapo: Spatial Association of Polygon Types}
\description{
The \code{sapo} package provides functions, with methods to
sumarise and visualize hypothesis tests for spatial association of
different types of polygons spatial association.
}
\section{tpsa functions and methods}{

\itemize{
\item psam_mc
\item gof_mc
\item summary
\item plot
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lcgodoy/sapo/}
  \item Report bugs at \url{https://github.com/lcgodoy/sapo/issues/}
}

}
\author{
\strong{Maintainer}: Lucas da Cunha Godoy \email{lcgodoy@duck.com} (\href{https://orcid.org/0000-0003-4265-972X}{ORCID})

}
