% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tshift.R
\name{translate_by_pt}
\alias{translate_by_pt}
\title{Translate an \code{sf} object by a "point"}
\usage{
translate_by_pt(pt, poly)
}
\arguments{
\item{pt}{\code{sfc} representing a shift.}

\item{poly}{\code{sfc} of \code{sf} to be shifted}
}
\value{
a \code{sf} or \code{sfc} representing \code{poly} shifted by \code{pt}
}
\description{
Translate an \code{sf} object by a "point"
}
\author{
Lucas Godoy
}
