% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_h}
\alias{calc_h}
\title{\eqn{h_{12}(t)}}
\usage{
calc_h(x, method = "min", L = FALSE, dists = NULL)
}
\arguments{
\item{x}{distance matrix}

\item{method}{method to deal with broken polygons}

\item{L}{logical scalar indicating if the L function should be used instead}

\item{dists}{vector of distances to compute \eqn{h_{12}(t)}.}
}
\value{
a numeric vector
}
\description{
Computes the \eqn{h_{12}} (K or L) based on a distance matrix based on a method
}
