% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_functions.R
\name{mc_iterations}
\alias{mc_iterations}
\title{psat_mc2 - auxiliar function}
\usage{
mc_iterations(obj1_shift, obj_sp2, niter)
}
\arguments{
\item{obj1_shift}{object from class \code{SpatialPolygons} or \code{SpatialPoints}}

\item{obj_sp2}{object from class \code{SpatialPolygons} or \code{SpatialPoints}}

\item{niter}{\code{numeric} value indicating number of iterations}
}
\value{
a \code{numeric vector}.
}
\description{
Internal use.
}
