% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{poly_rf2}
\alias{poly_rf2}
\title{Polygons' Random Shift - 2}
\usage{
poly_rf2(objsp, bbox_max)
}
\arguments{
\item{objsp}{object from class \code{SpatialPolygons}}

\item{bbox_max}{Boundary box from class \code{matrix}}
}
\value{
an object from class \code{SpatialPolygons} randomly translated
}
\description{
Polygons' Random Shift - 2
}
