% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{psam}
\alias{psam}
\title{Polygons' Spatial Association Measure}
\usage{
psam(obj_sp1, obj_sp2, correction = "none", ...)
}
\arguments{
\item{obj_sp1}{a \code{SpatialPolygons} or \code{SpatialPointsDataFrame} object}

\item{obj_sp2}{a \code{SpatialPolygons} or \code{SpatialPointsDataFrame} object}

\item{correction}{a \code{character} giving the edge correction to be used. Possible
entries are \code{c('none', 'torus', 'guard')}.}

\item{...}{parameters associated with edge correction.}
}
\value{
a \code{numeric} \code{scalar} with the polygons' spatial
 association measure for the corresponding objects.
}
\description{
test statistic for the MC test.
}
