% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psat.R
\docType{package}
\name{tpsa}
\alias{tpsa}
\alias{tpsa-package}
\title{tpsa: A package for testing polygons spatial association.}
\description{
The \code{tpsa} package provides functions, with methods to sumarise and
visualize it, for polygons spatial association test.
}
\section{tpsa functions and methods}{

\itemize{
  \item poly_mc
  \item summary
  \item plot
}
}

